backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
numpy
pypdf
python-docx
chromadb
google-generativeai